MC-40 Refactored password component
