+++ src/pages/components/index.html
@@ -17,19 +17,11 @@
 
     <div class="row password input">
         <p>Password</p>
-        <div id="pswd" class="container-fluid" >
-            <form action="" class="form-control"></form>
-                <div class="form-group pswd-box">
-                    <i class="fa-solid fa-lock lock-icon"></i>
-                    <label for="control-label" class="pswd-label">Senha</label>
-                    <input type="password" class="form-control pswd-input" placeholder="Senha">
-                    <i class="fa-solid fa-eye-slash eye-icon"></i>
-                    <span></span>
-                </div>
-                <div class="form-group pswd-box">
-                    <input type="button" value="Entrar">
-                </div> 
-            </form>
+        <div id="pswd" class="container-fluid pswd-box" >
+            <i class="fa-solid fa-lock lock-icon"></i>
+            <label for="control-label" class="pswd-label">Senha</label>
+            <input type="password" class="form-control pswd-input" placeholder="Senha">
+            <i class="fa-solid fa-eye-slash eye-icon"></i>
         </div>
     </div>
