+++ src/pages/components/main.js
@@ -1,9 +1,9 @@
 import { LayoutProviderService } from '../../services/layout-provider.service';
-import PasswordInput from '../../components/password-input.components';
+import PasswordInput from '../../components/password.components';
 
 import './style.scss';
 import page from './index.html?raw';
 
 LayoutProviderService.inject(page).then(() => {
-    const PasswordInputComponents = new PasswordInput('#pswd');
+    const password = new PasswordInput('#pswd');
 });
